backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
passlib==1.7.4
bcrypt==4.1.1
python-jose==3.3.0
cryptography==41.0.7